chore: moved homepageStats and LiveAndUpcoming getter hooks to state
