+++ src/pages/index.tsx
@@ -9,9 +9,9 @@ import { dehydrate, QueryClient } from '@tanstack/react-query'
 
 // SSR getters
 import { getIndustryStats } from 'state/lhd/hooks/useGetIndustryStats'
-import { getHomepageStats } from 'hooks/queries/useGetHomepageStats'
+import { getHomepageStats } from 'state/homepage/hooks/useGetHomepageStats'
 import { getTvlStats } from 'state/homepage/hooks/useGetTvlStats'
-import { getLiveAndUpcoming } from 'hooks/queries/useGetLiveAndUpcoming'
+import { getLiveAndUpcoming } from 'state/homepage/hooks/useGetLiveAndUpcoming'
 
 // Constants
 import { QUERY_KEYS } from 'config/constants/queryKeys'
